Move get_normals() to graphics::utils from forms
This funciton is more generally useful
